Split character builder component modules
